Fix test issues caused by merging
