+++ platforms/jvm/jvm-services/src/test/groovy/org/gradle/internal/jvm/inspection/DefaultJvmMetadataDetectorTest.groovy
@@ -136,7 +136,7 @@ class DefaultJvmMetadataDetectorTest extends Specification {
         def metadata = detector.getMetadata(testLocation(javaHome))
 
         then:
-        assert metadata.hasCapability(JavaInstallationCapability.J9_VIRTUAL_MACHINE) == isJ9
+        assert metadata.capabilities.contains(JavaInstallationCapability.J9_VIRTUAL_MACHINE) == isJ9
 
         where:
         jdk              | systemProperties         | isJ9
